Renamed electrolux_group integration to electrolux
