+++ homeassistant/components/aeg/manifest.json
@@ -1,6 +1,6 @@
 {
-    "domain": "aeg",
-    "name": "AEG",
-    "integration_type": "virtual",
-    "supported_by": "electrolux_group"
+  "domain": "aeg",
+  "name": "AEG",
+  "integration_type": "virtual",
+  "supported_by": "electrolux"
 }
