+++ homeassistant/components/electrolux/const.py
@@ -1,8 +1,8 @@
-"""Constants for Electrolux Group integration."""
+"""Constants for Electrolux integration."""
 
 from homeassistant.const import __version__ as HA_VERSION
 
-DOMAIN = "electrolux_group"
+DOMAIN = "electrolux"
 
 CONF_REFRESH_TOKEN = "refresh_token"
