+++ homeassistant/components/electrolux/coordinator.py
