+++ homeassistant/components/electrolux/entity_helper.py
