+++ homeassistant/components/electrolux/quality_scale.yaml
