+++ homeassistant/components/electrolux/strings.json
@@ -1,11 +1,11 @@
 {
   "config": {
     "abort": {
+      "added_to_existing_entry": "The appliance is added.",
       "already_configured": "This Electrolux account is already configured.",
-      "wrong_account": "The account is wrong.",
       "already_in_progress": "[%key:common::config_flow::abort::already_in_progress%]",
       "missing_id": "The appliance id is missing.",
-      "added_to_existing_entry": "The appliance is added."
+      "wrong_account": "The account is wrong."
     },
     "error": {
       "auth_failed": "Authentication failed. Please check your credentials.",
@@ -13,6 +13,10 @@
       "unknown": "[%key:common::config_flow::error::unknown%]"
     },
     "step": {
+      "discovery_confirm": {
+        "description": "A new appliance \"{device}\" has been discovered. Do you want to add it to Home Assistant?",
+        "title": "New {brand} appliance found"
+      },
       "user": {
         "data": {
           "access_token": "[%key:common::config_flow::data::access_token%]",
@@ -24,43 +28,39 @@
           "api_key": "Your Electrolux Group for Developer API key.",
           "refresh_token": "The refresh token used to renew your access token."
         },
-        "title": "Configure your Electrolux Group account",
-        "description": "Please [click here]({portal_link}) to generate new access and refresh tokens, then paste them below."
-      },
-      "discovery_confirm": {
-        "title": "New {brand} appliance found",
-        "description": "A new appliance \"{device}\" has been discovered. Do you want to add it to Home Assistant?"
+        "description": "Please [click here]({portal_link}) to generate new access and refresh tokens, then paste them below.",
+        "title": "Configure your Electrolux Group account"
       }
     }
   },
   "entity": {
     "sensor": {
-      "connection_state": {
-        "name": "Connection state"
-      },
-      "start_at": {
-        "name": "Start at"
-      },
-      "running_time": {
-        "name": "Running time"
-      },
       "appliance_state": {
         "name": "Appliance state"
       },
-      "food_probe_state": {
-        "name": "Food probe state"
+      "connection_state": {
+        "name": "Connection state"
+      },
+      "display_temperature": {
+        "name": "Current temperature"
       },
       "door_state": {
         "name": "Door state"
       },
-      "remote_control": {
-        "name": "Remote control"
+      "food_probe_state": {
+        "name": "Food probe state"
       },
       "food_probe_temperature": {
         "name": "Food probe temperature"
       },
-      "display_temperature": {
-        "name": "Current temperature"
+      "remote_control": {
+        "name": "Remote control"
+      },
+      "running_time": {
+        "name": "Running time"
+      },
+      "start_at": {
+        "name": "Start at"
       }
     }
   }
