+++ homeassistant/components/plushome/manifest.json
@@ -1,6 +1,6 @@
 {
-    "domain": "plushome",
-    "name": "+Home",
-    "integration_type": "virtual",
-    "supported_by": "electrolux_group"
+  "domain": "plushome",
+  "name": "+Home",
+  "integration_type": "virtual",
+  "supported_by": "electrolux"
 }
