+++ requirements_test_all.txt
@@ -766,7 +766,7 @@ ekey-bionyxpy==1.0.1
 # homeassistant.components.electric_kiwi
 electrickiwi-api==0.9.14
 
-# homeassistant.components.electrolux_group
+# homeassistant.components.electrolux
 electrolux-group-developer-sdk==0.1.3
 
 # homeassistant.components.elevenlabs
