+++ tests/components/electrolux/test_config_flow.py
@@ -10,7 +10,7 @@ from electrolux_group_developer_sdk.client.failed_connection_exception import (
 )
 
 from homeassistant import config_entries, data_entry_flow
-from homeassistant.components.electrolux_group.const import CONF_REFRESH_TOKEN, DOMAIN
+from homeassistant.components.electrolux.const import CONF_REFRESH_TOKEN, DOMAIN
 from homeassistant.const import CONF_ACCESS_TOKEN, CONF_API_KEY
 from homeassistant.core import HomeAssistant
 
@@ -19,10 +19,10 @@ async def test_user_flow_success(hass: HomeAssistant) -> None:
     """Test a successful user config flow."""
     with (
         patch(
-            "homeassistant.components.electrolux_group.config_flow.TokenManager"
+            "homeassistant.components.electrolux.config_flow.TokenManager"
         ) as mock_token_manager,
         patch(
-            "homeassistant.components.electrolux_group.config_flow.ApplianceClient"
+            "homeassistant.components.electrolux.config_flow.ApplianceClient"
         ) as mock_appliance_client,
     ):
         mock_token_manager_instance = mock_token_manager.return_value
@@ -36,8 +36,8 @@ async def test_user_flow_success(hass: HomeAssistant) -> None:
             DOMAIN, context={"source": config_entries.SOURCE_USER}
         )
 
-        assert result["type"] == "form"
-        assert result["errors"] == {}
+        assert result.get("type") == "form"
+        assert result.get("errors") == {}
 
         user_input = {
             CONF_ACCESS_TOKEN: "test_access_token",
@@ -49,9 +49,9 @@ async def test_user_flow_success(hass: HomeAssistant) -> None:
             result["flow_id"], user_input=user_input
         )
 
-        assert result2["type"] == "create_entry"
-        assert result2["title"] == "Electrolux"
-        assert result2["data"] == user_input
+        assert result2.get("type") == "create_entry"
+        assert result2.get("title") == "Electrolux"
+        assert result2.get("data") == user_input
 
 
 async def test_user_flow_invalid_auth(hass: HomeAssistant) -> None:
@@ -61,10 +61,10 @@ async def test_user_flow_invalid_auth(hass: HomeAssistant) -> None:
         DOMAIN, context={"source": config_entries.SOURCE_USER}
     )
 
-    assert result["type"] == "form"
+    assert result.get("type") == "form"
 
     with patch(
-        "homeassistant.components.electrolux_group.config_flow.TokenManager"
+        "homeassistant.components.electrolux.config_flow.TokenManager"
     ) as mock_token_manager:
         mock_token_manager_instance = mock_token_manager.return_value
         mock_token_manager_instance.ensure_credentials.side_effect = (
@@ -80,8 +80,8 @@ async def test_user_flow_invalid_auth(hass: HomeAssistant) -> None:
             },
         )
 
-    assert result2["type"] == data_entry_flow.FlowResultType.FORM
-    assert result2["errors"] == {"base": "auth_failed"}
+    assert result2.get("type") == data_entry_flow.FlowResultType.FORM
+    assert result2.get("errors") == {"base": "auth_failed"}
 
 
 async def test_user_flow_failed_connection(hass: HomeAssistant) -> None:
@@ -91,10 +91,10 @@ async def test_user_flow_failed_connection(hass: HomeAssistant) -> None:
         DOMAIN, context={"source": config_entries.SOURCE_USER}
     )
 
-    assert result["type"] == "form"
+    assert result.get("type") == "form"
 
     with patch(
-        "homeassistant.components.electrolux_group.config_flow.ApplianceClient"
+        "homeassistant.components.electrolux.config_flow.ApplianceClient"
     ) as mock_appliance_client:
         mock_appliance_client_instance = AsyncMock()
         mock_appliance_client_instance = mock_appliance_client.return_value
@@ -111,8 +111,8 @@ async def test_user_flow_failed_connection(hass: HomeAssistant) -> None:
             },
         )
 
-    assert result2["type"] == data_entry_flow.FlowResultType.FORM
-    assert result2["errors"] == {"base": "cannot_connect"}
+    assert result2.get("type") == data_entry_flow.FlowResultType.FORM
+    assert result2.get("errors") == {"base": "cannot_connect"}
 
 
 async def test_user_flow_unknown_error(hass: HomeAssistant) -> None:
@@ -122,10 +122,10 @@ async def test_user_flow_unknown_error(hass: HomeAssistant) -> None:
         DOMAIN, context={"source": config_entries.SOURCE_USER}
     )
 
-    assert result["type"] == "form"
+    assert result.get("type") == "form"
 
     with patch(
-        "homeassistant.components.electrolux_group.config_flow.ApplianceClient"
+        "homeassistant.components.electrolux.config_flow.ApplianceClient"
     ) as mock_appliance_client:
         mock_appliance_client_instance = AsyncMock()
         mock_appliance_client_instance = mock_appliance_client.return_value
@@ -140,8 +140,8 @@ async def test_user_flow_unknown_error(hass: HomeAssistant) -> None:
             },
         )
 
-    assert result2["type"] == data_entry_flow.FlowResultType.FORM
-    assert result2["errors"] == {"base": "unknown"}
+    assert result2.get("type") == data_entry_flow.FlowResultType.FORM
+    assert result2.get("errors") == {"base": "unknown"}
 
 
 async def test_user_flow_abort_flow(hass: HomeAssistant) -> None:
@@ -151,11 +151,11 @@ async def test_user_flow_abort_flow(hass: HomeAssistant) -> None:
         DOMAIN, context={"source": config_entries.SOURCE_USER}
     )
 
-    assert result["type"] == "form"
+    assert result.get("type") == "form"
 
     # Patch _authenticate_user to raise AbortFlow
     with patch(
-        "homeassistant.components.electrolux_group.config_flow.ElectroluxConfigFlow._authenticate_user",
+        "homeassistant.components.electrolux.config_flow.ElectroluxConfigFlow._authenticate_user",
         side_effect=data_entry_flow.AbortFlow("already_in_progress"),
     ):
         result2 = await hass.config_entries.flow.async_configure(
@@ -167,5 +167,5 @@ async def test_user_flow_abort_flow(hass: HomeAssistant) -> None:
             },
         )
 
-    assert result2["type"] == "abort"
-    assert result2["reason"] == "already_in_progress"
+    assert result2.get("type") == "abort"
+    assert result2.get("reason") == "already_in_progress"
